move useLiveKitRoom hook
